numDays: long -> Integer

+++ src/main/java/org/mskcc/limsrest/controller/GetSequencingRequests.java
@@ -17,7 +17,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static java.lang.Long.parseLong;
+import static com.velox.api.datarecord.DataFieldHelper.parseInteger;
 import static org.mskcc.limsrest.util.Utils.getResponseEntity;
 
 @RestController
@@ -51,9 +51,9 @@ public class GetSequencingRequests {
 
         Map<String, Object> resp = new HashMap<>();
 
-        Long numDays = 0L;
+        Integer numDays = 0;
         try {
-            numDays = parseLong(days);
+            numDays = parseInteger(days);
         } catch (NumberFormatException e) {
             String clientMessage = String.format("Couldn't process input days: %s", days);
             log.error(String.format("%s. Error: %s", clientMessage, e.getMessage()));
